project/SKYNET/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
websockets==12.0
aiofiles==23.2.1
pydantic==2.5.0
python-socketio==5.10.0
alembic==1.13.0
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
aiohttp==3.9.1
psutil==5.9.6
cryptography==41.0.7
google-generativeai==0.3.1
openai==1.6.1
anthropic==0.8.1
supabase==2.3.4